优化we_chat_zombie_friends的脚本文件更新方式

+++ we_chat_zombie_friends/utils/update_util.js
@@ -31,23 +31,39 @@ module.exports = (() => {
         let download_all_file = true;
         if (checkUpdate()) {
             let remote_project = readRemoteProject();
-            for (let i = 0, count = 1; i < remote_project["files"].length; i++) {
-                let file_path = remote_project["files"][i];
-                let result = download(base_url + file_path);
-                if (result["success"]) {
-                    dialog.progress = 100 * count++ / remote_project["files"].length;
-                    files.createWithDirs(".download_files/" + file_path);
-                    files.write(".download_files/" + file_path, result["content"]);
-                } else {
-                    download_all_file = false;
+            if (remote_project != undefined) {
+                let local_project = JSON.parse(files.read("project.json"));
+                let remote_files_md5 = remote_project["files_md5"];
+                let local_files_md5 = local_project["files_md5"];
+                let diff_file_count = 0;
+                for (let key in remote_files_md5) {
+                    if (local_files_md5[key] == undefined || local_files_md5[key] != remote_files_md5[key]) {
+                        diff_file_count++;
+                    }
                 }
-            }
-            if (download_all_file) {
-                for (let i = 0; i < remote_project["files"].length; i++) {
-                    let file_path = remote_project["files"][i];
-                    files.copy(".download_files/" + file_path, file_path);
+                let count = 1;
+                for (let key in remote_files_md5) {
+                    if (local_files_md5[key] == undefined || local_files_md5[key] != remote_files_md5[key]) {
+                        let result = download(base_url + key);
+                        if (result["success"]) {
+                            dialog.progress = 100 * count++ / diff_file_count;
+                            files.createWithDirs(".download_files/" + key);
+                            files.write(".download_files/" + key, result["content"]);
+                        } else {
+                            download_all_file = false;
+                        }
+                    }
                 }
-                files.removeDir(".download_files/");
+                if (download_all_file) {
+                    for (let key in remote_files_md5) {
+                        if (local_files_md5[key] == undefined || local_files_md5[key] != remote_files_md5[key]) {
+                            files.copy(".download_files/" + key, key);
+                        }
+                    }
+                    files.removeDir(".download_files/");
+                }
+            } else {
+                download_all_file = false;
             }
         }
         return download_all_file;
